Add support for Require commands (#113)
Hello!

I'm very happy to see this project as I personally think it is very
important for Coq to have a good code formatter. The progress so far is
quite promising and I would love to help in any way I can. This PR
implements decent but not complete support for `Require` vernacular
commands. Please let me know if there is anything you object to with
regards to code style or other issues due to not being completely
familiar with the codebase yet.

I added some tests but they are rather superficial as there isn't much
formatting to be done for `Require` commands per se.

P.S.: When I opened this PR I didn't have commit signing, that's why
there are a few extra force pushes while I was setting it up. Everything
should be okay now.

---------

Signed-off-by: Master Builder <gylfi@stebbi.is>